requirements.txt
fastapi
uvicorn[standard]
requests
pillow
openai
python-dotenv
pydantic
pydantic_extra_types
sqlalchemy
python-multipart
passlib
pyjwt
bcrypt
httpx
pytest